+++ node/resolvers/document/index.ts
@@ -1,54 +1,104 @@
-import { prop, union } from 'ramda'
-import {parseFieldsToJson }  from '../../utils'
-import { mapKeyValues } from '../../utils/object'
+import { compose, map, union, prop, replace } from 'ramda'
+import { parseFieldsToJson } from '../../utils/object'
+
 
 export const queries = {
-  documents: async (_: any, args: any, { dataSources: { document } }: any) => {
+
+  documents: async (_: any, args: DocumentsArgs, context: Context) => {
     const { acronym, fields, page, pageSize, where, schema } = args
+    const { clients: { masterdata } } = context
     const fieldsWithId = union(fields, ['id'])
     const data = schema
-      ? await document.searchDocumentsWithSchema(
-          acronym,
-          fieldsWithId,
-          where,
-          schema,
-          { page, pageSize }
-        )
-      : await document.searchDocuments(acronym, fieldsWithId, where, {
-          page,
-          pageSize,
-        })
-    return data.map((doc: any) => ({
-      cacheId: doc.id,
-      fields: mapKeyValues(doc),
-      id: doc.id,
-    }))
+    ? await masterdata.searchDocumentsWithSchema(
+        acronym,
+        fieldsWithId,
+        where,
+        schema,
+        { page, pageSize }
+      )
+    : await masterdata.searchDocuments(acronym, fieldsWithId, where, {
+      page,
+      pageSize,
+    })
+    return map(
+      (document: any) =>
+      ({
+        cacheId: document.id,
+        id: document.id,
+        fields: mapKeyAndStringifiedValues(document)
+      })
+    )(data)
   },
 
-  document: async (_: any, args: any, { dataSources: { document } }: any) => {
+  document: async (_: any, args: DocumentArgs, context: Context) => {
     const { acronym, fields, id } = args
-    const data = await document.getDocument(acronym, id, fields)
-    return { id, cacheId: id, fields: mapKeyValues(data) }
-  },
+    const { clients: { masterdata } } = context
+    const data = await masterdata.getDocument(acronym, id, fields)
+    return {
+      cacheId: id,
+      id,
+      fields: mapKeyAndStringifiedValues(data)}
+  }
 }
 
 export const mutations = {
-  createDocument: async (_: any, args: any, { dataSources: { document } }: any) => {
+
+  createDocument: async (_: any, args: CreateDocumentArgs, context: Context) => {
     const { acronym, document: { fields } } = args
-    const { Id, Href, DocumentId } = await document.createDocument(acronym, fields)
-    return { cacheId: DocumentId, id: Id, href: Href, documentId: DocumentId }
+    const { clients: { masterdata } } = context
+    const response = await masterdata.createDocument(acronym, parseFieldsToJson(fields))
+    const documentId = removeAcronymFromId(acronym, response)
+    return {
+      cacheId: documentId,
+      id: prop('Id', response),
+      href: prop('Href', response),
+      documentId: removeAcronymFromId(acronym, response),
+    }
   },
 
-  updateDocument: async (_: any, args: any, { dataSources: { document } }: any) => {
+  updateDocument: async (_: any, args: UpdateDocumentArgs, context: Context) => {
     const { acronym, document: { fields } } = args
-    const id = prop('id', parseFieldsToJson(fields))
-    const { Id, Href, DocumentId } = await document.updateDocument(acronym, id, fields)
-    return { cacheId: DocumentId, id: Id, href: Href, documentId: DocumentId }
+    const documentId = prop('id', parseFieldsToJson(fields)) as string
+    const { clients: { masterdata }, vtex: { account } } = context
+    await masterdata.updateDocument(acronym, documentId, parseFieldsToJson(fields))
+    return {
+      cacheId: documentId,
+      documentId,
+      href: generateHref(account, acronym, documentId),
+      id: getId(acronym, documentId),
+    }
   },
 
-  deleteDocument: async (_: any, args: any, { dataSources: { document } }: any) => {
+  deleteDocument: async (_: any, args: DeleteDocumentArgs, context: Context) => {
     const { acronym, documentId } = args
-    await document.deleteDocument(acronym, documentId)
-    return { id: documentId }
+    const { clients: { masterdata }, vtex: { account } } = context
+    await masterdata.deleteDocument(acronym, documentId)
+    return {
+      documentId,
+      href: generateHref(account, acronym, documentId),
+      id: getId(acronym, documentId),
+      cacheId: documentId,
+    }
   }
 }
+
+/**
+ * Map a document object to a list of {key: 'property', value: 'propertyValue'},
+ * Uses `JSON.stringify` in every value.
+ */
+const mapKeyAndStringifiedValues = (document: any) => Object.keys(document).map(key => ({
+  key,
+  value: typeof(document[key]) === 'string' ? document[key] : JSON.stringify(document[key])
+}))
+
+const removeAcronymFromId = (acronym: string, data: { Id: string }) => {
+  return compose<any, any, any>(
+    replace(new RegExp(`${acronym}-`), ''),
+    prop('Id')
+  )(data)
+}
+
+const getId = (acronym: string, documentId: string) => `${acronym}-${documentId}`
+
+const generateHref = (account: string, acronym: string, documentId: string) =>
+  `http://api.vtex.com/${account}/dataentities/${acronym}/documents/${documentId}`