+++ node/resolvers/profile/index.ts
@@ -9,19 +9,22 @@ import {
 } from './services'
 
 export const mutations = {
-  createAddress: (_: any, { fields }: any, context: Context) => createAddress(context, fields),
+  createAddress: (_: any, { fields }: any, context: Context) =>
+    createAddress(context, fields),
 
-  deleteAddress: (_: any, { id }: any, context: Context) => deleteAddress(context, id),
+  deleteAddress: (_: any, { id }: any, context: Context) =>
+    deleteAddress(context, id),
 
-  updateAddress: (_: any, args: any, context: Context) => updateAddress(context, args),
+  updateAddress: (_: any, args: any, context: Context) =>
+    updateAddress(context, args),
 
   updateProfile: (_: any, { fields, customFields }: any, context: Context) =>
     updateProfile(context, fields, customFields),
 
-  updateProfilePicture: (_: any, { file }: any, context: Context) =>
+  updateProfilePicture: (_: any, { file }: { file: any }, context: Context) =>
     updateProfilePicture(context, file),
 
-  uploadProfilePicture: (_: any, { file }: any, context: Context) =>
+  uploadProfilePicture: (_: any, { file }: { file: any }, context: Context) =>
     updateProfilePicture(context, file),
 
   subscribeNewsletter: async (_: any, { email }: any, context: Context) => {
@@ -32,11 +35,12 @@ export const mutations = {
     })
 
     return true
-  }
+  },
 }
 
 export const queries = {
-  profile: (_: any, { customFields }: any, context: any) => getProfile(context, customFields),
+  profile: (_: any, { customFields }: any, context: any) =>
+    getProfile(context, customFields),
 }
 
 export const fieldResolvers = fieldR