+++ node/utils/cookie.ts
@@ -1,14 +1,37 @@
-import cookies from 'cookie'
-
-const appendToCookie = (ctx: Context, extraValue: string) => {
-  const {request: { headers: { cookie } }} = ctx
-  ctx.request.headers.cookie = `${cookie}; ${extraValue}`
-}
+import * as Cookies from "cookies"
+import { parse } from 'cookie'
+import { keys } from 'ramda'
 
 const isUserLoggedIn = (ctx: Context) => {
-  const { vtex: { account }, headers: {cookie} } = ctx
-  const parsedCookie = cookies.parse(cookie)
-  return !!parsedCookie[`VtexIdclientAutCookie_${account}`]
+  const { vtex: { account } } = ctx
+  return !!ctx.cookies.get(`VtexIdclientAutCookie_${account}`)
 }
 
-export { appendToCookie, isUserLoggedIn }
+const parseCookie = (cookie: string) => {
+  const parsed = parse(cookie)
+  const cookieName = keys(parsed)[0]
+  const cookieValue = parsed[cookieName]
+
+  const extraOptions = {
+    path: parsed.path,
+    domain: parsed.domain,
+    expires: parsed.expires ? new Date(parsed.expires) : undefined,
+  }
+  return {
+    name: cookieName,
+    value: cookieValue,
+    options: extraOptions,
+  }
+}
+
+/** Checkout cookie methods */
+const CHECKOUT_COOKIE = 'checkout.vtex.com'
+
+const checkoutCookieFormat = (orderFormId: string) => `${CHECKOUT_COOKIE}=__ofid=${orderFormId};`
+
+const getOrderFormIdFromCookie = (cookies: Cookies) => {
+  const cookie = cookies.get(CHECKOUT_COOKIE)
+  return cookie && cookie.split('=')[1]
+}
+
+export { isUserLoggedIn, CHECKOUT_COOKIE, checkoutCookieFormat, getOrderFormIdFromCookie, parseCookie }