+++ CHANGELOG.md
@@ -7,35 +7,88 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
-## [2.73.3] - 2019-05-14
+## [2.75.0] - 2019-05-20
+### Changed
+- Use better resolver architecture at productSearch query.
+
+### Added
+- breadcrum resolver at productSearch.
+
+## [2.74.4] - 2019-05-20
+
 ### Fixed
+
+- **_updateProfilePicture_** mutation.
+
+## [2.74.3] - 2019-05-17
+
+## [2.74.2] - 2019-05-16
+
+### Fixed
+
+- Fix `orders` and `shipping` routes in Checkout Client.
+
+## [2.74.1] - 2019-05-16
+
+## Added
+
+- Check when user is impersonated in `getPasswordLastUpdate`
+
+## [2.74.0] - 2019-05-16
+
+### Fixed
+
+- `document` resolvers working according to the graphql schema.
+
+### Changed
+
+- Ported DocumentDataSource to MasterData IOClient
+- Ported CheckoutDataSource to Checkout IOClient
+
+## [2.73.3] - 2019-05-14
+
+### Fixed
+
 - Use filter for availability on facets query as well.
 
 ## [2.73.2] - 2019-05-14
+
 ### Fixed
+
 - Search metadata for Brands.
 
 ## [2.73.1] - 2019-05-13
+
 ### Fixed
+
 - Adds missing facets fields back after schema breaking change
 
 ## [2.73.0] - 2019-05-13
+
 ### Changed
+
 - Translate terms (with `Messages`) before sending them to search api for `productSearch` and `facets` resolvers
 
 ## [2.72.2] - 2019-05-10
+
 ### Fixed
+
 - Remove comma from invalid characters of product search.
 
 ## [2.72.1] - 2019-05-09
+
 ### Fixed
+
 - Remove slugify on facets when computing selected property.
 
 ## [2.72.0] - 2019-05-09
+
 ### Added
+
 - Autocomplete field resolvers for internationalization
 
 ### Changed
+
 - scope from SEGMENT to PUBLIC for Brand related queries
 - Splits Deparment, Brand etc types from Facets
 - Product id as translation provider id
@@ -51,6 +104,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 
 - Slugify facets when checking if is selected.
+  ]
 
 ## [2.71.1] - 2019-05-07